loading/lazyload/move-element-and-scroll.tentative.html

<!DOCTYPE html>
<head>
  <title>Images with loading='lazy' load being moved to another document and then scrolled to.</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="common.js"></script>
</head>

<!--
Marked as tentative until https://github.com/whatwg/html/pull/3752 is landed.
-->

<body>
  <img id="in_viewport" src='resources/image.png'>
  <div style="height:10000px;"></div>
  <img id="below_viewport" src='resources/image.png?second' loading="lazy">

  <script>
    const in_viewport_element = document.getElementById("in_viewport");
    const below_viewport_element = document.getElementById("below_viewport");

    async_test(function(t) {
      var iframe = document.createElement('iframe');
      iframe.srcdoc = "<body></body>";
      iframe.onload = function() {
        var adopted_img = iframe.contentDocument.adoptNode(below_viewport_element);
        iframe.contentDocument.body.appendChild(adopted_img);
        window.scrollTo(0, 10000);
        t.step_timeout(function() { assert_false(is_image_fully_loaded(below_viewport_element)); t.done(); }, 2000);
      };
      document.body.insertBefore(iframe, in_viewport_element);
    }, "Test that <img> below viewport is not loaded when moved to another document and then scrolled to.");
</script>
</body>
